Navbar changes in mobile view.
The navbar has a button toggler, but I decided to change it to a comment (aka taking it out of my page) in order to have the logo and site title in the center. It is a cool feature, however I personally prefer to see the links when I view a site on mobile.